refactor(tarko): simplify code editor components
Remove motion animations and duplicate code
Consolidate header logic into shared component
