+++ multimodal/tarko/agent-web-ui/src/sdk/code-editor/index.ts
@@ -1,3 +1,4 @@
 export { CodeEditor } from './CodeEditor';
 export { MonacoCodeEditor } from './MonacoCodeEditor';
 export { DiffViewer } from './DiffViewer';
+export { CodeEditorHeader } from './CodeEditorHeader';
